replaced ToolTip with HoverToolTip and then renamed HoverToolTip to ToolTip, removed tech debt of old ToolTip
