--- src/shared/ToolTip/IconWithTooltip.module.scss
@@ -1,40 +1,0 @@
-.tooltip {
-  position: relative;
-  top: -6em;
-  left: -10em;
-}
-
-.tooltiptext {
-  box-sizing: border-box;
-  background-color: white;
-  font-size: 1.3rem;
-  width: 250px;
-  border-radius: 6px;
-  padding: 0.5rem 1rem;
-  position: absolute;
-
-  border: 1px solid black;
-}
-
-.arrow {
-  height: 10px;
-  width: 10px;
-  background: white;
-  top: 3.85em;
-  left: 10.1em;
-  position: absolute;
-  z-index: 1;
-  transform: rotate(45deg);
-  border-right: 1px solid black;
-  border-bottom: 1px solid black;
-}
-
-@media only screen and (max-width: 481px) {
-  .arrow {
-    top: 3.95em;
-  }
-}
-
-.color_blue_link {
-  color: #205493;
-}
