+++ src/shared/ToolTip/ToolTip.module.scss
@@ -4,7 +4,7 @@
 
     .tooltipTextTop {
       position: absolute;
-      top: calc(-100% - 60px);
+      top: calc(-100% - 65px);
       transform: translateX(-50%);
       background-color: rgba($color: #000000, $alpha: .8);
       color: white;
@@ -12,10 +12,12 @@
       border-radius: 4px;
       height: 80px; // setting hard width and height to keep positioning constant
       width: 200px;
-      text-align: center;
       white-space: normal;
       overflow: auto; // sets a scrollbar
       z-index: 1; // Ensure the tooltip is above other elements
+      display: flex;
+      align-items: center;
+      justify-content: center;
   }
 
     .tooltipTextRight {
@@ -33,6 +35,9 @@
         white-space: normal;
         overflow: auto;
         z-index: 1;
+        display: flex;
+        align-items: center;
+        justify-content: center;
     }
 
     .tooltipTextLeft {
@@ -50,6 +55,9 @@
         white-space: normal;
         overflow: auto;
         z-index: 1;
+        display: flex;
+        align-items: center;
+        justify-content: center;
     }
 
     .tooltipTextBottom {
@@ -67,5 +75,8 @@
         white-space: normal;
         overflow: auto;
         z-index: 1;
+        display: flex;
+        align-items: center;
+        justify-content: center;
     }
 }
