--- src/shared/ToolTip/index.css
@@ -1,48 +1,0 @@
-.tooltip {
-  position: relative;
-}
-
-.tooltip .tooltiptext {
-  box-sizing: border-box;
-  background-color: #4d4d4d;
-  font-size: 0.9rem;
-  line-height: 1rem;
-  color: #fff;
-  text-align: center;
-  border-radius: 6px;
-  padding: 0.5rem;
-  position: absolute;
-  z-index: 1;
-  top: 175%;
-  left: 50%;
-}
-
-.tooltip .tooltiptext-small {
-  width: 6rem;
-  margin-left: -3rem;
-}
-
-.tooltip .tooltiptext-medium {
-  width: 10rem;
-  margin-left: -5rem;
-}
-
-.tooltip .tooltiptext-large {
-  width: 16rem;
-  margin-left: -8rem;
-}
-
-.tooltip .tooltiptext::after {
-  content: '';
-  position: absolute;
-  bottom: 100%;
-  left: 50%;
-  margin-left: -5px;
-  border-width: 5px;
-  border-style: solid;
-  border-color: transparent transparent #4d4d4d transparent;
-}
-
-.tooltip:hover .tooltiptext {
-  visibility: visible;
-}
